requirements.txt
Flask==2.3.3
Pillow==10.0.0
opencv-python-headless==4.8.0.74
numpy==1.24.3
rembg==2.0.47
gunicorn==21.2.0